+++ wasm/lib/src/module.dart
@@ -20,6 +20,10 @@ class WasmModule {
     _module = runtime.compile(this, data);
   }
 
+  /// Asynchronously compile a module.
+  static Future<WasmModule> compileAsync(Uint8List data) async =>
+      Future<WasmModule>(() => WasmModule(data));
+
   /// Returns a [WasmInstanceBuilder] that is used to add all the imports that
   /// the module needs before instantiating it.
   WasmInstanceBuilder builder() => WasmInstanceBuilder._(this);
@@ -211,6 +215,10 @@ class WasmInstanceBuilder {
     }
     return WasmInstance._(_module, _importOwner, _imports, _wasiEnv);
   }
+
+  /// Asynchronously build the module instance.
+  Future<WasmInstance> buildAsync() async =>
+      Future<WasmInstance>(() => build());
 }
 
 // TODO: should not be required once the min supported Dart SDK includes