+++ wasm/CHANGELOG.md
@@ -4,6 +4,8 @@
 - All WASM modules and isntances use a singleton store, to enable sharing of
   memory and functions.
 - Add options to setup.dart for configuring the build.
+- Add `WasmModule.compileAsync` and `WasmInstanceBuilder.buildAsync`, so that
+  the web API can be supported in the future.
 
 ## 0.1.0+1